+++ CHANGELOG.md
@@ -13,11 +13,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [0.0.37] - 2025-08-18
+
+### Added
+- add twilio to notification chanel
+
+### Changed
+
+### Fixed
+- add the server name in message for pushover
+- fix broken FindAll with tags
+
+- Fix custom domain issues
+
 ## [0.0.36] - 2025-07-28
 
 ### Added
 
-- implement rendering certificate information for https monitors 
+- implement rendering certificate information for https monitors
 
 ### Changed